Mais alguns controllers configurados(com relacionamento)

+++ src/main/java/com/store/storeapp/service/CustomerService.java
@@ -1,6 +1,7 @@
 package com.store.storeapp.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,15 +19,29 @@ public class CustomerService {
 		return customerRepository.findAll();
 	}
 	
+	public Optional<Customer> getOneCustomer(Long id){
+		return customerRepository.findById(id);
+	}
+	
 	public void setCustomer(Customer customer) {
 		customerRepository.save(customer);
 	}
 	
-	public void updateCustomer(Customer id) {
-		customerRepository.save(id);
+	public Optional<Customer> updateCustomer(Customer updateCustomer, Long id) {
+		Optional<Customer> customer = getOneCustomer(id);
+		if(customer.isPresent()) {
+			return customer = Optional.of(customerRepository.save(updateCustomer));
+		}
+		return null;
+		
 	}
 	
-	public void deleteCustomer(Customer id) {
-		customerRepository.delete(id);
+	public boolean deleteCustomer(Customer id) {
+		List<Customer> customer = getAllCustomer();
+		if(customer.contains(id)) {
+			customerRepository.delete(id);
+			return true;
+		}
+		return false;			
 	}
 }
